move pages under /[network]
