+++ packages/web/src/pages/[network]/liquidity/[version].tsx
@@ -5,17 +5,17 @@ import { Footer } from 'src/components/organisms/Footer'
 import { Container } from 'src/components/atoms/Container'
 import { Headline } from 'src/components/atoms/Headline'
 import { H2 } from 'src/components/atoms/Typography'
-import { Nav } from '../../components/_pages/liquidity/Nav'
+import { Nav } from '../../../components/_pages/liquidity/Nav'
 import styled from 'styled-components'
-import { Deposit } from '../../components/_pages/liquidity/Deposit'
-import { Withdraw } from '../../components/_pages/liquidity/Withdraw'
+import { Deposit } from '../../../components/_pages/liquidity/Deposit'
+import { Withdraw } from '../../../components/_pages/liquidity/Withdraw'
 import { Stats } from 'src/components/_pages/liquidity/Stats'
 import { Informations } from 'src/components/_pages/liquidity/Informations'
-import { VersionSwitching } from '../../components/_pages/liquidity/VersionSwitching'
+import { VersionSwitching } from '../../../components/_pages/liquidity/VersionSwitching'
 import {
   GEYSER_V1_ETHDEV_V2_ADDRESS,
   GEYSER_V2_ETHDEV_V2_ADDRESS
-} from '../../fixtures/_pages/liquidity/constants/address'
+} from '../../../fixtures/_pages/liquidity/constants/address'
 import { useRouter } from 'next/router'
 // import Error from 'next/error'
 import { getPath } from 'src/fixtures/utility/route'
