+++ packages/web/src/pages/[network]/positions/[stokenId]/index.tsx
