+++ packages/web/src/pages/[network]/stats/index.tsx
